unit Testing Group Velocity Implementation
Update with files concerning testing of the implementation of group velocity so that the wakefields may travel at a speed other than the speed of the probe. Debug objects are used to get field data within the simulation through the wakefield for visualization and testing purposes. Debug mode can be activated using input file and is currently supported only for single particles. Testing for Group Velocity can be found in ipynb files